add connection handler for handling actor connections
